0.9.0

+++ src/main/java/org/cardanofoundation/explorer/api/model/response/script/nativescript/NativeScriptResponse.java
@@ -1,11 +1,11 @@
- package org.cardanofoundation.explorer.api.model.response.script.nativescript;
+package org.cardanofoundation.explorer.api.model.response.script.nativescript;
 
-import org.cardanofoundation.ledgersync.common.common.nativescript.ScriptType;
-import lombok.*;
-
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
+
+import lombok.*;
+
+import com.bloxbean.cardano.client.transaction.spec.script.ScriptType;
 
 @Getter
 @Setter
@@ -17,11 +17,11 @@ public class NativeScriptResponse {
   private Long numberOfTokens;
   private Long numberOfAssetHolders;
   private ScriptType conditionType;
-  private BigInteger required;
+  private Integer required;
   private List<String> keyHashes;
   private LocalDateTime after;
   private LocalDateTime before;
-  private List <String> associatedAddress;
+  private List<String> associatedAddress;
   private String script;
   private Boolean isOneTimeMint;
   private Boolean verifiedContract;
